requirements.txt
Flask
msal
python-dotenv
requests
requests-toolbelt